chore(app): moved useAttestation to separate file
